gaggl_like_worker/src/requirements.txt
flask==1.0.0
requests==2.18.4
python-dotenv==0.12.0
redis==3.5.3